+++ src/state/scene/manager/id_manager.rs
@@ -1,11 +1,11 @@
 pub struct IdManager
 {
-    material_id: u32,
-    texture_id: u32,
-    node_id: u32,
-    instance_id: u32,
-    camera_id: u32,
-    light_id: u32,
+    material_id: u64,
+    texture_id: u64,
+    node_id: u64,
+    instance_id: u64,
+    camera_id: u64,
+    light_id: u64,
 }
 
 impl IdManager
@@ -23,42 +23,42 @@ impl IdManager
         }
     }
 
-    pub fn get_next_material_id(&mut self) -> u32
+    pub fn get_next_material_id(&mut self) -> u64
     {
         self.material_id = self.material_id + 1;
 
         self.material_id
     }
 
-    pub fn get_next_texture_id(&mut self) -> u32
+    pub fn get_next_texture_id(&mut self) -> u64
     {
         self.texture_id = self.texture_id + 1;
 
         self.texture_id
     }
 
-    pub fn get_next_node_id(&mut self) -> u32
+    pub fn get_next_node_id(&mut self) -> u64
     {
         self.node_id = self.node_id + 1;
 
         self.node_id
     }
 
-    pub fn get_next_instance_id(&mut self) -> u32
+    pub fn get_next_instance_id(&mut self) -> u64
     {
         self.instance_id = self.instance_id + 1;
 
         self.instance_id
     }
 
-    pub fn get_next_camera_id(&mut self) -> u32
+    pub fn get_next_camera_id(&mut self) -> u64
     {
         self.camera_id = self.camera_id + 1;
 
         self.camera_id
     }
 
-    pub fn get_next_light_id(&mut self) -> u32
+    pub fn get_next_light_id(&mut self) -> u64
     {
         self.light_id = self.light_id + 1;
